+++ example/lib/screens/navigation/navigation_view.dart
@@ -164,34 +164,31 @@ class NavigationViewPage extends ScrollablePage {
                 ),
               ],
             ),
-            // content: NavigationBody(
-            //   index: topIndex,
-            //   transitionBuilder: pageTransition == 'default'
-            //       ? null
-            //       : (child, animation) {
-            //           switch (pageTransition) {
-            //             case 'entrance':
-            //               return EntrancePageTransition(
-            //                 child: child,
-            //                 animation: animation,
-            //               );
-            //             case 'drill in':
-            //               return DrillInPageTransition(
-            //                 child: child,
-            //                 animation: animation,
-            //               );
-            //             case 'horizontal':
-            //               return HorizontalSlidePageTransition(
-            //                 child: child,
-            //                 animation: animation,
-            //               );
-            //             default:
-            //               throw UnsupportedError(
-            //                 '$pageTransition is not a supported transition',
-            //               );
-            //           }
-            //         },
-            // ),
+            transitionBuilder: pageTransition == 'default'
+                ? null
+                : (child, animation) {
+                    switch (pageTransition) {
+                      case 'entrance':
+                        return EntrancePageTransition(
+                          child: child,
+                          animation: animation,
+                        );
+                      case 'drill in':
+                        return DrillInPageTransition(
+                          child: child,
+                          animation: animation,
+                        );
+                      case 'horizontal':
+                        return HorizontalSlidePageTransition(
+                          child: child,
+                          animation: animation,
+                        );
+                      default:
+                        throw UnsupportedError(
+                          '$pageTransition is not a supported transition',
+                        );
+                    }
+                  },
           ),
         ),
         codeSnippet: '''NavigationView(
@@ -206,37 +203,33 @@ class NavigationViewPage extends ScrollablePage {
       PaneItem(
         icon: const Icon(FluentIcons.home),
         title: const Text('Home'),
+        body: BodyItem(),
       ),
       PaneItem(
         icon: const Icon(FluentIcons.issue_tracking),
         title: const Text('Track an order'),
         infoBadge: const InfoBadge(source: Text('8')),
+        body: BodyItem(),
       ),
       PaneItemExpander(
         icon: const Icon(FluentIcons.account_management),
         title: const Text('Account'),
+        body: BodyItem(),
         items: [
           PaneItem(
             icon: const Icon(FluentIcons.mail),
             title: const Text('Mail'),
+            body: BodyItem(),
           ),
           PaneItem(
             icon: const Icon(FluentIcons.calendar),
             title: const Text('Calendar'),
+            body: BodyItem(),
           ),
         ],
       ),
     ],
   ),
-  body: NavigationBody(
-    index: topIndex,
-    children: const [
-      BodyItem(),
-      BodyItem(),
-      BodyItem(),
-      BodyItem(),
-    ],
-  )
 )''',
       ),
     ];