+++ lib/src/controls/navigation/navigation_view/view.dart
@@ -23,8 +23,6 @@ const double _kDefaultAppBarHeight = 50.0;
 ///
 /// See also:
 ///
-///   * [NavigationBody], a widget that implement fluent transitions into
-///     [NavigationView]
 ///   * [NavigationPane], the pane used by [NavigationView], that can be
 ///     displayed either at the left and top
 ///   * [TabView], a widget similar to [NavigationView], useful to display
@@ -70,11 +68,9 @@ class NavigationView extends StatefulWidget {
   /// [EntrancePageTransition] is used.
   ///
   /// ```dart
-  /// NavigationBody(
-  ///   transitionBuilder: (child, animation) {
-  ///     return DrillInPageTransition(child: child, animation: animation);
-  ///   },
-  /// ),
+  /// transitionBuilder: (child, animation) {
+  ///   return DrillInPageTransition(child: child, animation: animation);
+  /// },
   /// ```
   ///
   /// See also:
@@ -115,7 +111,7 @@ class NavigationViewState extends State<NavigationView> {
   ///
   /// It's also used to display and control the [Scrollbar] introduced
   /// by the panes.
-  late ScrollController scrollController;
+  late final ScrollController paneScrollController;
 
   /// The key used to animate between open and compact display mode
   final _panelKey = GlobalKey();
@@ -125,13 +121,10 @@ class NavigationViewState extends State<NavigationView> {
 
   final Map<int, GlobalKey> _itemKeys = {};
 
-  /// The overlay entry used for minimal pane
-  OverlayEntry? minimalOverlayEntry;
-
   bool _minimalPaneOpen = false;
   late bool _compactOverlayOpen;
 
-  int oldIndex = 0;
+  int _oldIndex = 0;
 
   PaneDisplayMode? _autoDisplayMode;
 
@@ -148,15 +141,14 @@ class NavigationViewState extends State<NavigationView> {
   @override
   void initState() {
     super.initState();
-    scrollController = ScrollController(
-      debugLabel: '${widget.runtimeType} scroll controller',
-      keepScrollOffset: true,
-    );
-    scrollController.addListener(() {
-      if (mounted) setState(() {});
-    });
+    paneScrollController = widget.pane?.scrollController ??
+        ScrollController(
+          debugLabel: '${widget.runtimeType} scroll controller',
+          keepScrollOffset: true,
+        );
+    paneScrollController.addListener(_handleScrollControllerEvent);
 
-    generateKeys();
+    _generateKeys();
 
     _compactOverlayOpen = PageStorage.of(context)?.readState(
           context,
@@ -165,26 +157,31 @@ class NavigationViewState extends State<NavigationView> {
         false;
   }
 
+  void _handleScrollControllerEvent() {
+    if (mounted) setState(() {});
+  }
+
   @override
   void didUpdateWidget(NavigationView oldWidget) {
     super.didUpdateWidget(oldWidget);
-    if (widget.pane?.scrollController != scrollController) {
-      scrollController = widget.pane?.scrollController ?? scrollController;
+    if (widget.pane?.scrollController != paneScrollController) {
+      paneScrollController =
+          widget.pane?.scrollController ?? paneScrollController;
     }
 
     if (oldWidget.pane?.selected != widget.pane?.selected) {
-      oldIndex = oldWidget.pane?.selected ?? -1;
+      _oldIndex = oldWidget.pane?.selected ?? -1;
     }
 
     if (oldWidget.pane?.effectiveItems.length !=
         widget.pane?.effectiveItems.length) {
       if (widget.pane?.effectiveItems.length != null) {
-        generateKeys();
+        _generateKeys();
       }
     }
   }
 
-  void generateKeys() {
+  void _generateKeys() {
     if (widget.pane == null) return;
     _itemKeys
       ..clear()
@@ -201,7 +198,12 @@ class NavigationViewState extends State<NavigationView> {
 
   @override
   void dispose() {
-    scrollController.dispose();
+    // If the controller was created locally, dispose it
+    if (widget.pane?.scrollController == null) {
+      paneScrollController.dispose();
+    } else {
+      paneScrollController.removeListener(_handleScrollControllerEvent);
+    }
     super.dispose();
   }
 
@@ -209,15 +211,17 @@ class NavigationViewState extends State<NavigationView> {
   Widget build(BuildContext context) {
     assert(debugCheckHasFluentTheme(context));
     assert(debugCheckHasFluentLocalizations(context));
-    assert(debugCheckHasDirectionality(context));
     assert(debugCheckHasMediaQuery(context));
+    assert(debugCheckHasDirectionality(context));
 
     final Brightness brightness = FluentTheme.of(context).brightness;
     final NavigationPaneThemeData theme = NavigationPaneTheme.of(context);
-    final localizations = FluentLocalizations.of(context);
-    final appBarPadding =
-        EdgeInsets.only(top: widget.appBar?.finalHeight(context) ?? 0.0);
-    final direction = Directionality.of(context);
+    final FluentLocalizations localizations = FluentLocalizations.of(context);
+    final MediaQueryData mediaQuery = MediaQuery.of(context);
+    final EdgeInsetsGeometry appBarPadding = EdgeInsetsDirectional.only(
+      top: widget.appBar?.finalHeight(context) ?? 0.0,
+    );
+    final TextDirection direction = Directionality.of(context);
 
     Color? overlayBackgroundColor() {
       if (theme.backgroundColor == null) {
@@ -247,7 +251,7 @@ class NavigationViewState extends State<NavigationView> {
         /// (641px to 1007px).
         /// - Only a menu button (minimal) on small window widths (640px or less).
         double width = consts.biggest.width;
-        if (width.isInfinite) width = MediaQuery.of(context).size.width;
+        if (width.isInfinite) width = mediaQuery.size.width;
 
         if (width <= 640) {
           _autoDisplayMode = PaneDisplayMode.minimal;
@@ -314,7 +318,7 @@ class NavigationViewState extends State<NavigationView> {
                   content: ClipRect(child: body),
                   listKey: _listKey,
                   paneKey: _panelKey,
-                  scrollController: scrollController,
+                  scrollController: paneScrollController,
                   pane: pane,
                 ),
               ),
@@ -549,7 +553,7 @@ class NavigationViewState extends State<NavigationView> {
                   curve: theme.animationCurve ?? Curves.linear,
                   start: _minimalPaneOpen ? 0.0 : -kOpenNavigationPaneWidth,
                   width: kOpenNavigationPaneWidth,
-                  height: MediaQuery.of(context).size.height,
+                  height: mediaQuery.size.height,
                   child: PaneScrollConfiguration(
                     child: ColoredBox(
                       color: Colors.black,
@@ -599,7 +603,7 @@ class NavigationViewState extends State<NavigationView> {
           displayMode: _compactOverlayOpen ? PaneDisplayMode.open : displayMode,
           minimalPaneOpen: _minimalPaneOpen,
           pane: widget.pane,
-          oldIndex: oldIndex,
+          oldIndex: _oldIndex,
           child: PaneItemKeys(keys: _itemKeys, child: paneResult),
         ),
       );
@@ -608,13 +612,15 @@ class NavigationViewState extends State<NavigationView> {
 
   // ignore: non_constant_identifier_names
   Widget PaneScrollConfiguration({required Widget child}) {
-    return PrimaryScrollController(
-      controller: scrollController,
-      child: ScrollConfiguration(
-        behavior: const NavigationViewScrollBehavior(),
-        child: child,
-      ),
-    );
+    return Builder(builder: (context) {
+      return PrimaryScrollController(
+        controller: paneScrollController,
+        child: ScrollConfiguration(
+          behavior: const NavigationViewScrollBehavior(),
+          child: child,
+        ),
+      );
+    });
   }
 }
 
@@ -738,7 +744,6 @@ class NavigationAppBar with Diagnosticable {
   double finalHeight(BuildContext context) {
     assert(debugCheckHasMediaQuery(context));
     final mediaQuery = MediaQuery.of(context);
-
     final topPadding = mediaQuery.viewPadding.top;
 
     return height + topPadding;